marketplace_project/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.1
dj-database-url==2.1.0
Pillow==10.2.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1 